Delete old Controller V2
